+++ apps/client/src/components/common/FeaturesSection.tsx
