+++ apps/client/src/components/common/Footer.tsx
