+++ apps/client/src/components/common/FullScreenModal.tsx
