+++ apps/client/src/components/common/TrustedBySection.tsx
