+++ apps/client/src/components/core/ProjectsSection.tsx
