+++ apps/client/src/components/forms/AddTaskModal.tsx
@@ -3,14 +3,18 @@ import { Dialog } from "@headlessui/react";
 import axios from "axios";
 import { useAuth } from "@clerk/clerk-react";
 import { motion } from "framer-motion";
-import { getApiEndpoint } from "../utils/api";
-import { TaskStatus, Priority, TaskType, User } from "../utils/taskTypes";
-import { Calendar } from "./ui/calendar";
-import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover";
-import { Button } from "./ui/button";
+import { getApiEndpoint } from "@/lib/api";
+import { TaskStatus, Priority, TaskType, User } from "@/lib/types/taskTypes";
+import { Calendar } from "@/components/ui/calendar";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover";
+import { Button } from "@/components/ui/button";
 import { format } from "date-fns";
 import { CalendarIcon } from "lucide-react";
-import { cn } from "../lib/utils";
+import { cn } from "@/lib/utils";
 
 interface AddTaskModalProps {
   isOpen: boolean;
