+++ apps/client/src/components/forms/SprintFormModal.tsx
@@ -3,12 +3,16 @@ import { motion, AnimatePresence } from "framer-motion";
 import { X, CalendarIcon } from "lucide-react";
 import { useAuth } from "@clerk/clerk-react";
 import axios from "axios";
-import api from "../utils/api";
-import { Calendar } from "./ui/calendar";
-import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover";
-import { Button } from "./ui/button";
+import api from "@/lib/api";
+import { Calendar } from "@/components/ui/calendar";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@/components/ui/popover";
+import { Button } from "@/components/ui/button";
 import { format } from "date-fns";
-import { cn } from "../lib/utils";
+import { cn } from "@/lib/utils";
 
 interface Project {
   id: string;
