+++ apps/client/src/components/icons/DiscordIcon.tsx
