+++ apps/client/src/components/icons/PlatFormMockup.tsx
