+++ apps/client/src/main.tsx
@@ -1,7 +1,8 @@
 // import { StrictMode } from 'react'
 import { createRoot } from "react-dom/client";
 import { ClerkProvider } from "@clerk/clerk-react";
-import { QueryClientProvider, queryClient } from "./utils/queryClient";
+import { QueryClientProvider } from "@tanstack/react-query";
+import { queryClient } from "./lib/api/queryClient";
 import "./index.css";
 import App from "./App.tsx";
